Refactor: BottomSheet 컴포넌트 Compound Component 패턴으로 리팩토링
