+++ src/app/collection/page.css.ts
@@ -124,3 +124,14 @@ export const addFolderButton = style({
   fontSize: '1.6rem',
   fontWeight: 700,
 });
+
+// BottomSheet Input
+export const contentInput = style({
+  padding: '2rem 2.4rem',
+  backgroundColor: '#F5F6FA',
+  borderRadius: 18,
+
+  color: vars.color.black,
+  fontSize: '1.6rem',
+  fontWeight: 400,
+});
